Set python3 as default, if it exists.
If there is a python3 installed on the system, then it is going to be
the default to be used. Otherwise, we use just the python command. It
can even be a python3 version, in systems that only have python3 or the
python command is linked to the python3 version.

This commit fixes #896.