Fix PathFinder.find_module AttributeError for Python 3.12
PathFinder.find_module() has been deprecated since Python 3.4 in favor of
find_spec(), and it's finally removed in Python 3.12. This line will
throw an AttributeError which makes pymode completely unusable with
python 3.12.

It was a hacky workaround introduced in #1028. Maybe we can completely
remove this workaround because it's 4 years ago and the minimum
supported python version is now 3.6+.